# perf: use global prefix slices to avoid allocations